Grafos con matriz de adyacencia

+++ src/mincut/IMinCut.java
@@ -3,5 +3,5 @@ package mincut;
 import graph.Graph;
 
 public interface IMinCut {
-  double minCut(Graph graph);
+  int minCut(Graph graph);
 }
